requirements.txt
fastapi
uvicorn[standard]
openai==0.28
python-dotenv
python-multipart
torch
torchaudio
requests
